+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Your contribution here.
 * [#30](https://github.com/ruby-grape/grape-swagger-entity/pull/30): Fix nested exposures with an alias - [@Kukunin](https://github.com/Kukunin).
+* [#31](https://github.com/ruby-grape/grape-swagger-entity/pull/31): Respect `required: true` for nested attributes as well - [@Kukunin](https://github.com/Kukunin).
 
 ### 0.2.2 (November 3, 2017)